include sha256 field in jf rt s results.
